+++ NEWS.md
@@ -3,6 +3,7 @@
 - In the gwstates and cowstates data, the "iso3c" columns have been renamed to "gwn" and "cown", as they were not in fact ISO 3-character country abbreviations (#19).
 - plot_missing()'s "space" argument has been renamed "ccode". 
 - id_date_sequence() now works with integer years, there is no need to convert them to Date class anymore (#20).
+- fixed a mispelled country name, "Rumania" to "Romania".
 
 ## Improved support for state_panel shortcuts